переделал Дз по модульности
