--- js/create-ordered-id-generator.js
@@ -1,10 +1,0 @@
-const createOrderedIdGenerator = () => { // функция генерации последовательных id
-  let lastOrderedId = 0;
-
-  return function () {
-    lastOrderedId += 1;
-    return lastOrderedId;
-  };
-}
-
-export {createOrderedIdGenerator};
